+++ packages/web/app/utils/saveExcelFile.js
@@ -1,5 +1,5 @@
 import XLSX from 'xlsx';
-import { sanitizeFileName } from './sanitizeFileName';
+import { createFileSystemHandle } from './fileSystemAccess';
 
 const stringifyIfNonDateObject = val =>
   typeof val === 'object' && !(val instanceof Date) && val !== null ? JSON.stringify(val) : val;
@@ -17,22 +17,9 @@ export async function saveExcelFile({ data, metadata, defaultFileName = '', book
   XLSX.utils.book_append_sheet(book, dataSheet, 'report');
   XLSX.utils.book_append_sheet(book, metadataSheet, 'metadata');
 
-  const types = [];
-  if (bookType === 'xlsx') {
-    types.push({
-      accept: { 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet': ['.xlsx'] },
-    });
-  }
-
-  if (bookType === 'csv') {
-    types.push({
-      accept: { 'text/csv': ['.csv'] },
-    });
-  }
-
-  const fileHandle = await window.showSaveFilePicker({
-    suggestedName: sanitizeFileName(`${defaultFileName}`),
-    types,
+  const fileHandle = await createFileSystemHandle({
+    defaultFileName,
+    extensions: [bookType],
   });
 
   const writable = await fileHandle.createWritable();
